--- display_welcome_message.py
@@ -1,14 +1,0 @@
-""" Display a welcome message """
-
-from random import shuffle
-import argparse
-import itertools
-import sys
-
-
-def display_welcome_message():
-    print()
-    print("GatorGrouper: Automatically Assign Students to Groups")
-    print("https://github.com/gkapfham/gatorgrouper")
-    print()
-
