+++ parse_arguments.py
@@ -1,9 +1,6 @@
 """ Parses the arguments provided on the command-line """
 
-from random import shuffle
 import argparse
-import itertools
-import sys
 import logging
 import grouping_method
 
@@ -11,7 +8,7 @@ import grouping_method
 from defaults import *
 from read_student_file import read_student_file
 
-def parse_gatorgrouper_arguments(args):
+def parse_arguments(args):
 
     gg_parser = argparse.ArgumentParser(
         formatter_class=argparse.ArgumentDefaultsHelpFormatter)
@@ -45,7 +42,8 @@ def parse_gatorgrouper_arguments(args):
     gg_parser.add_argument(
         "--random",
         help="Use random grouping method",
-        action="store_const", dest="grouping_method", const=grouping_method.RANDOM, default=grouping_method.RANDOM
+        action="store_const", dest="grouping_method",
+        const=grouping_method.RANDOM, default=grouping_method.RANDOM
     )
 
     gg_parser.add_argument(
@@ -68,9 +66,12 @@ def parse_gatorgrouper_arguments(args):
 
     gg_arguments_finished = gg_parser.parse_args(args)
 
-    logging.basicConfig(format="%(levelname)s:%(pathname)s: %(message)s", level=gg_arguments_finished.logging_level)
+    logging.basicConfig(format="%(levelname)s:%(pathname)s: %(message)s",
+                        level=gg_arguments_finished.logging_level)
 
-    if check_valid_group_size(gg_arguments_finished.group_size, read_student_file(gg_arguments_finished.students_file)) == False:
+    if not check_valid_group_size(
+            gg_arguments_finished.group_size,
+            read_student_file(gg_arguments_finished.students_file)):
         quit()
 
     return gg_arguments_finished
@@ -78,7 +79,8 @@ def parse_gatorgrouper_arguments(args):
 def check_valid_group_size(group_size, students_list):
     students_list_length = len(students_list)
     if (group_size <= 1 or group_size > students_list_length / 2): # indicates invalid group size
-        logging.error("Group size: " + str(group_size) + "\nNumber of students: " + str(students_list_length) + "\nGroup size must be greater than 1 and less than or equal to half of the number of students.")
+        logging.error("Group size: " + str(group_size) + "\nNumber of students: " + str(students_list_length) +
+                "\nGroup size must be greater than 1 and less than or equal to half of the number of students.")
         return False
     else:
         logging.info("Group size: " + str(group_size) + "\nNumber of students: " + str(students_list_length) + "\nValid group size.")
